Brush up app module and whole sequence #12
* SizeCalculator: Device impl class and interface.
